requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
cairosvg==2.7.1
reportlab==4.2.2
qrcode[pil]==8.0
Pillow==11.0.0
